requirements.txt

Django==5.1.4
djangorestframework
djangorestframework-simplejwt